coda/requirements.txt
attrs==19.3.0
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
codaio==0.5.0
decorator==4.4.2
envparse==0.2.0
idna==2.10
inflection==0.3.1
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
todoist-python==8.1.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.5
